Add files via upload
Added more flags to determine what seq gen to do
Added fast dinucleotide seq gen

+++ fimo_scanner.py
@@ -12,74 +12,41 @@ import pandas as pd
 import numpy as np
 
 ######################################### Sequence Generator Main #########################################
-def run_fimo_scanner(outdir, sample, cpus, motifs, 
-                     threshold_fimo, background_file, genome, 
-                     verbose, experimental_fimo, whole_genome_fimo):
+def run_fimo_scanner(verbose, outdir, sample, cpus, motifs, 
+                     threshold_fimo, background_file, seq_type):
+    
+    fimo_dirs(verbose, outdir=outdir, seq_type=seq_type)
     if verbose == True: 
         print("---------Calculating GC Content of Motifs from Motif Database (.meme file)----------")
     motif_list = fimo_motif_names(verbose=verbose,motifs=motifs)
     get_gc(verbose, outdir=outdir, sample=sample, motifs=motifs, motif_list=motif_list, alphabet=['A','C','G','T'])
-        
+    
     if verbose == True: 
-        print("---------FIMO Scan: Identifying Locations of Motif Hits in Simulated Genome----------")
+        print("---------FIMO Scan: Identifying Locations of Motif Hits -------------")
+        print("This scan is for the " + seq_type + " genome.")
         print('Initializing ' + str(cpus) + ' cpus to run FIMO scan.')
         print('Start time: %s' % str(datetime.datetime.now()))
         start_time = int(time.time())
-    fimo_dirs(verbose, outdir=outdir, seq_type='simulated')
-    seq_type='simulated'
+    seq_type=seq_type
     pool = multiprocessing.Pool(cpus)
     results = pool.map(partial(scanner, 
-                               inputs=[verbose, outdir, sample, motifs, threshold_fimo, background_file, genome, seq_type]), motif_list)
-    if verbose == True: 
-        stop_time = int(time.time())
-        print('Stop time: %s' % str(datetime.datetime.now()))
-        print("Total Run time :", (stop_time-start_time)/3600, " hrs")
+                               inputs=[verbose, outdir, sample, motifs, threshold_fimo, background_file, seq_type]), motif_list)
+    if verbose == True:
+        print('FIMO scan complete at: %s' % str(datetime.datetime.now()))
         print("----------------Convert Simulated FIMO to Bed Format-----------------------")
-    fimotobed(verbose, outdir, seq_type='simulated')
+    fimotobed(verbose, outdir, seq_type=seq_type)
     if verbose == True: 
         print("Bed file columns: ['chr','start', 'stop', 'motif_id','score', 'strand','identifier', 'motif_region_name']")
+        stop_time = int(time.time())
+        print("The scan for the " + seq_type + " genome is complete.")
+        print('Stop time: %s' % str(datetime.datetime.now()))
+        print("Total Run time :", (stop_time-start_time)/3600, " hrs")
 
-    if experimental_fimo == True:
-        if verbose == True: 
-            print("---------FIMO Scan: Identifying Locations of Motif Hits in Experimental Genome----------")
-            print('Start time: %s' % str(datetime.datetime.now()))
-            start_time = int(time.time())
-        fimo_dirs(verbose, outdir=outdir, seq_type='experimental')
-        seq_type='experimental'
-        pool.map(partial(scanner,
-                         inputs=[verbose, outdir, sample, motifs, threshold_fimo, background_file, genome, seq_type]), motif_list)
-        if verbose == True: 
-            stop_time = int(time.time())
-            print('Stop time: %s' % str(datetime.datetime.now()))
-            print("Total Run time :", (stop_time-start_time)/3600, " hrs")
-            print("----------------Convert Experimental FIMO to Bed Format-----------------------")
-        fimotobed(verbose, outdir, seq_type='experimental')
-    
-    elif whole_genome_fimo == True:
-        if verbose == True: 
-            print("---------FIMO Scan: Identifying Locations of Motif Hits in Whole Genome----------")
-            print('Start time: %s' % str(datetime.datetime.now()))
-            start_time = int(time.time())
-        fimo_dirs(verbose, outdir=outdir, seq_type='whole_genome')
-        seq_type='whole_genome'
-        pool.map(partial(scanner,
-                         inputs=[verbose, outdir, sample, motifs, threshold_fimo, background_file, genome, seq_type]), motif_list)
-        if verbose == True: 
-            stop_time = int(time.time())
-            print('Stop time: %s' % str(datetime.datetime.now()))
-            print("Total Run time :", (stop_time-start_time)/3600, " hrs")
-            print("----------------Convert Whole Genome FIMO to Bed Format-----------------------")
-        fimotobed(verbose, outdir, seq_type='whole_genome')
-    else:
-        if verbose == True:
-            print('No experimental motifs scanned.')
-            print('Note: It is important that both the experimental motifs and simulated motifs are called with the same fimo parameters (meme file, cutoff and background file).')
-        
 ######################################### FIMO Functions #########################################
 
 def fimo_dirs(verbose, outdir, seq_type):
     if (path.exists(outdir + '/temp') == False):
-        os.system('mkdir -p ' + outdir + '/temp') 
+        os.system('mkdir -p ' + outdir + '/temp')   
     try:
         os.system('mkdir -p ' + outdir + '/temp/' + seq_type + '_fimo_out')
     except OSError:
@@ -87,7 +54,7 @@ def fimo_dirs(verbose, outdir, seq_type):
         sys.exit(1)
     else:
         if verbose == True:
-            print('Successfully created the directory %s' % outdir + '/temp/' + seq_type + '_fimo_out')
+            print(outdir + '/temp/' + seq_type + '_fimo_out exists.')
     
     if (path.exists(outdir + '/motifs') == False):
         os.system('mkdir -p ' + outdir + '/motifs')        
@@ -98,7 +65,7 @@ def fimo_dirs(verbose, outdir, seq_type):
         sys.exit(1)
     else:
         if verbose == True:
-            print('Successfully created the directory %s' % outdir + '/motifs/' + seq_type)
+            print(outdir + '/motifs/' + seq_type + ' exists.')
             
 def fimo_motif_names(verbose, motifs):
     '''Extracts motif names from a MEME formatted motif database'''
@@ -109,53 +76,53 @@ def fimo_motif_names(verbose, motifs):
                 line = line.strip('\n').split()
                 motif_name = line[-1]
                 motif_list.append(motif_name)
-    first = [motif_list[0:1]]
-    last =  [motif_list[-1:]]
     if verbose == True:
-        print('There are ' + str(len(motif_list)) + " motifs in this meme file. " + str(first) + '...' + str(last))
+        print('There are ' + str(len(motif_list)) + " motifs in this meme file.")
     return motif_list
 
 def get_gc(verbose, outdir, sample, motifs, motif_list, alphabet=['A','C','G','T']):
-    '''
-    Obtain a pssm model from a meme formatted database file.
-    '''
-    gc_out={}
-    for motif in motif_list:
-        motif_hit = False
-        PSSM = []
-        with open(motifs,'r') as F:
-            for line in F:
-                if 'MOTIF' in line:
-                    if motif in line:
-                        motif_hit = True
-                    else:
-                        motif_hit = False
-                elif motif_hit and 'URL' not in line and 'letter-probability' not in line and line != '\n':
-                    acgt_probabilities = [float(x) for x in line.strip('\n').split()]
-                    total_prob = sum(acgt_probabilities)
-                    acgt_probabilities = [x/total_prob for x in acgt_probabilities] #Convert to probabilities
-                    PSSM.append(acgt_probabilities)
+    '''Obtain a pssm model from a meme formatted database file.'''
+    if (path.exists(outdir + "/generated_sequences/" + str(sample) + "_motif_gc_percentage.txt") == False):
+        gc_out={}
+        for motif in motif_list:
+            motif_hit = False
+            PSSM = []
+            with open(motifs,'r') as F:
+                for line in F:
+                    if 'MOTIF' in line:
+                        if motif in line:
+                            motif_hit = True
+                        else:
+                            motif_hit = False
+                    elif motif_hit and 'URL' not in line and 'letter-probability' not in line and line != '\n':
+                        acgt_probabilities = [float(x) for x in line.strip('\n').split()]
+                        total_prob = sum(acgt_probabilities)
+                        acgt_probabilities = [x/total_prob for x in acgt_probabilities] #Convert to probabilities
+                        PSSM.append(acgt_probabilities)
 
-            gc = 0
-            for base in PSSM:
-                gc += base[alphabet.index('C')]
-                gc += base[alphabet.index('G')]
+                gc = 0
+                for base in PSSM:
+                    gc += base[alphabet.index('C')]
+                    gc += base[alphabet.index('G')]
 
-            gc = gc/float(len(PSSM))
+                gc = gc/float(len(PSSM))
+                if verbose == True:
+                    print('Percent GC content for ' + motif + ' is ' + str(gc))
+            gc_out[motif] = motif,gc,(len(PSSM))
+
+            df_gc_out = pd.DataFrame.from_dict(gc_out)
+            df_gc_out = df_gc_out.transpose()
+            df_gc_out.columns =['motif', 'percent_gc','motif_length']
+            df_gc_out.to_csv(outdir + "/generated_sequences/" + str(sample) + "_motif_gc_percentage.txt", header=None, index=False, sep='\t')
+        else:
             if verbose == True:
-                print('Percent GC content for ' + motif + ' is ' + str(gc))
-        gc_out[motif] = motif,gc
- 
-        df_gc_out = pd.DataFrame.from_dict(gc_out)
-        df_gc_out = df_gc_out.transpose()
-        df_gc_out.to_csv(outdir + "/generated_sequences/" + str(sample) + "_motif_gc_percentage.txt", header=None, index=False, sep='\t')
+                print('Motif length and GC content already calculated.')
+                print('File: ' + outdir + "/generated_sequences/" + str(sample) + "_motif_gc_percentage.txt")
 
 def scanner(motif_list, inputs):
-    verbose, outdir, sample, motifs, threshold_fimo, background_file, genome, seq_type = inputs
-    if seq_type == 'whole_genome':
-        fasta = genome
-    else:
-        fasta = outdir + '/generated_sequences/' + sample + '_' + seq_type + '.fa'
+    verbose, outdir, sample, motifs, threshold_fimo, background_file, seq_type = inputs
+    
+    fasta = outdir + '/generated_sequences/' + sample + '_' + seq_type + '.fa'
 
     if verbose == True:
         fimo_verbosity = '--verbosity 2 '
@@ -188,9 +155,10 @@ def fimotobed(verbose, outdir, seq_type):
                 df.drop(['count'], axis=1, inplace=True)
                 df = df[['sequence_name','start', 'stop', 'motif_id','score', 'strand','identifier', 'motif_region_name']]
                 df['start'] = df['start'].astype(int) 
-                df['stop'] = df['stop'].astype(int)                 
+                df['stop'] = df['stop'].astype(int)
+                df = df.sort_values(by=['chr', 'start'])
                 df.to_csv(outdir + '/motifs/' + seq_type + '/' + motif_name + '.sorted.bed', sep='\t', header=None, index=False)
                 
 def identifier(row):
-    ident= str(row['sequence_name']) + ':' + str(row['start']) + '-' + str(row['stop']) + '_' + str(row['strand'] + ';' + str(row['motif_id']))
+    ident= str(row['sequence_name']) + ':' + str(row['start']) + '-' + str(row['stop']) + '_' + str(row['strand'] + ';' + str(row['score']) + ';' + str(row['motif_id']))
     return ident
